requirements.txt
langchain>=0.1.0
langchain-google-genai>=0.0.5  # Use this instead of langchain-openai
langchain-community>=0.0.10
langchain-core>=0.1.10
google-generativeai>=0.3.0     # Gemini API library
beautifulsoup4>=4.12.2
httpx>=0.25.1
lxml>=4.9.3
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.4.2
python-dotenv>=1.0.0
